Fixes #26054 - VMware storage controller type chage
(cherry picked from commit 6223a9649b6955387c5661a0d23612c2556c75bc)
